+++ next-js/src/app/dashboard/expense/list-wrapper.jsx
@@ -4,10 +4,10 @@ import { ExpenseListView } from 'src/sections/expense/view';
 
 import { PageGuard } from 'src/auth/guard';
 
-export default function ExpenseListWrapper({ expenses, permissionError }) {
+export default function ExpenseListWrapper() {
   return (
     <PageGuard>
-      <ExpenseListView expenses={expenses} permissionError={permissionError} />
+      <ExpenseListView />
     </PageGuard>
   );
 }